Crate polkadot_gossip_support

This subsystem is responsible for keeping track of session changes and issuing a connection request to the relevant validators on every new session.

In addition to that, it creates a gossip overlay topology which limits the amount of messages sent and received to be an order of sqrt of the validators. Our neighbors in this graph will be forwarded to the network bridge with the NetworkBridgeRxMessage::NewGossipTopology message.

Structs

GossipSupport

The Gossip Support subsystem.